Routes creation and module aggregation
The module 'nodemon' was added, which serves to reload the server every
time a change is saved in the project files. The module is not essential
for the operation of the project, so when installing it, the command was
used:

```
npm i nodemon -D
```

There was a route that showed 'Index page', it was changed to an
'express' router. From 'express' the module 'Router' was called, this
module allows you to create an object that in turn allows you to define
'urls' or server paths. This object was stored in the 'router' constant.

The MVC model (model, view, controler) was used for route management, in
order to avoid handling the routes in a disorderly manner.

In the 'controlers' folder there are two files. One to store the
functions related to images called 'image.js', and another in charge of
the main page called 'home.js'.

The main functions of the routes definied in the file 'index.js' of the
'routes' folder were initialized in these files.