Upload images
In the form, which recived the image and its data, located in the
'index.js' file of the 'views' folder, a parameter was added indicating
that the form will recive heavy information, such as the image.

In the 'image.js' file o the 'controllers' file, the 'create' function
that recives the image and the data related to it is populated and the
processed and stored in the 'public/uploads' folder, with a random name
that it is imported from the file 'libs.js' in the folder 'helpers'.