Save image data in MON
To store an image, a model described in the 'image.js' file was created.
The models determine the parameters od the types of objects to be
stored, such as comments, images, users, among others.
To organixe the export of these from the 'models' folder, the 'index.js'
file that imports each model was created and exported as part of a
single object, which can be called from external files.